Move PropagationContext to Tracing
