+++ src/State/HubInterface.php
@@ -20,8 +20,6 @@ use Sentry\Tracing\TransactionContext;
  * This interface represent the class which is responsible for maintaining a
  * stack of pairs of clients and scopes. It is the main entry point to talk
  * with the Sentry client.
- *
- * @property \Sentry\State\PropagationContext $propagationContext
  */
 interface HubInterface
 {
