+++ tests/Tracing/DynamicSamplingContextTest.php
@@ -8,9 +8,9 @@ use PHPUnit\Framework\TestCase;
 use Sentry\ClientInterface;
 use Sentry\Options;
 use Sentry\State\Hub;
-use Sentry\State\PropagationContext;
 use Sentry\State\Scope;
 use Sentry\Tracing\DynamicSamplingContext;
+use Sentry\Tracing\PropagationContext;
 use Sentry\Tracing\TraceId;
 use Sentry\Tracing\Transaction;
 use Sentry\Tracing\TransactionContext;
